+++ partB/closed_event.h
@@ -4,15 +4,16 @@
 
 namespace mtm{
     class ClosedEvent : public BaseEvent{
-        PriorityQueue<long> event_invited;
+        PriorityQueue<int> event_invited;
     public:
         ClosedEvent() = default;
         ClosedEvent(DateWrap date, std::string name);
         ClosedEvent(const ClosedEvent& event);
         ~ClosedEvent() {}
-        void addInvitee(const long student);
-        void registerParticipant(const long student) override;
+        void addInvitee(const int student);
+        void registerParticipant(const int student) override;
         BaseEvent* clone() const override;
     };
 }
 #endif
+
